# MangaST_1_11/FrontMangaST/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - MKS",
    "description": "OpenApi documentation for Spring",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "MKS",
      "url": "https://mks.com/course",
      "email": "MKS@mail.ru"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082/mangaSt",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/user/updateProfileCover": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "loadNewProfileCover",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/mangaList/addToFavorite/{manga-id}": {
      "post": {
        "tags": [
          "manga-controller"
        ],
        "operationId": "addToFavorite",
        "parameters": [
          {
            "name": "manga-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/user/find/favoriteList": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "findUserFavoriteList",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseMangaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mangaList/find/{mangaPageId}": {
      "get": {
        "tags": [
          "manga-controller"
        ],
        "operationId": "findMangaByPageId",
        "parameters": [
          {
            "name": "mangaPageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MangaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mangaList/find/findTitle/{title}": {
      "get": {
        "tags": [
          "manga-controller"
        ],
        "operationId": "findByTitle",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "title",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseMangaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mangaList/find/findAllManga": {
      "get": {
        "tags": [
          "manga-controller"
        ],
        "operationId": "findAllManga",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseMangaResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/mangaList/favorite": {
      "get": {
        "tags": [
          "manga-controller"
        ],
        "operationId": "findAllFavoriteMangaByUser",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseMangaResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/mangaList/deleteFromFavorite/{manga-id}": {
      "delete": {
        "tags": [
          "manga-controller"
        ],
        "operationId": "deleteFromFavorite",
        "parameters": [
          {
            "name": "manga-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MangaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "mangaCover": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          },
          "achieved": {
            "type": "boolean"
          },
          "mangaPageId": {
            "type": "string"
          },
          "pathTitle": {
            "type": "string"
          },
          "typeManga": {
            "type": "string",
            "enum": [
              "MANGA",
              "MANHWA"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ANNOUNCE",
              "ONGOING",
              "PAUSED",
              "COMPLETED",
              "DISCONTINUED"
            ]
          },
          "description": {
            "type": "string"
          }
        }
      },
      "PageResponseMangaResponse": {
        "type": "object",
        "properties": {
          "listManga": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MangaResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}